completes server refactor
