+++ server/controllers/logout.js
@@ -1,0 +1,5 @@
+const logout = (req, res) => {
+    res.cookie('authToken', '').json('Logged out.');
+}
+
+module.exports = logout;
